ui: phase 1 — extract msgs and callbacks out of app.go
Pure mechanical split. Moves every *Msg type and every *Func callback
type out of internal/ui/app.go into two new files under the same
package. No semantic change: same declarations, same field names,
same documentation. All existing callers and tests continue to compile
and pass unchanged.

  internal/ui/msgs.go (new, 485 lines)
    - All exported *Msg types (ChannelSelectedMsg ... ToastMsg, plus
      the message-action family: MessageSentMsg, MessageSendFailedMsg,
      EditMessageMsg, DeleteMessageMsg, MessageEditedMsg, etc.)
    - All file-private message types: previewLoadedMsg, previewErrorMsg,
      previewSpinnerTickMsg, threadFetchDebounceMsg, autoScrollTickMsg,
      editEmptyToastMsg

  internal/ui/callbacks.go (new, 124 lines)
    - All *Func callback types injected via Set* methods on App
      (SwitchWorkspaceFunc, ChannelFetchFunc, MessageSendFunc,
      ThreadFetchFunc, ReactionAddFunc, PermalinkFetchFunc, etc.)
    - ChannelVisitRecorder, ChannelLookupFunc
    - clipboardReader type + defaultClipboardReader var

  internal/ui/app.go: 6216 -> 5656 lines (-560)
    560 deletions, 0 additions. App struct, NewApp, Init, Update,
    View, state types (Panel, navStack, editState, panelCache,
    dragState, loadingEntry, workspaceStatus) all stay put — those
    are targets for phases 2 and 4.

Verification: go vet clean, go build ./... clean, all 24 ui packages
test green (including the 35 Phase 0 characterization tests), View
benchmarks still execute.
