+++ internal/ui/app.go
@@ -98,278 +98,6 @@ const (
 	cacheFreshThreshold = 30 * time.Second
 )
 
-// Messages sent between components
-type (
-	ChannelSelectedMsg struct {
-		ID   string
-		Name string
-		// Type is the channel type ("channel", "private", "dm",
-		// "group_dm"); used to render a type-aware glyph in the
-		// message-pane header and status bar. May be empty when
-		// callers don't yet know the type — the UI then falls
-		// back to a default `#` glyph.
-		Type string
-		// FromHistory marks navigations synthesized by Ctrl+H /
-		// Ctrl+K. The case ChannelSelectedMsg handler suppresses
-		// pushing onto navHistory when this is true so back/forward
-		// walks don't grow the stack on every step. Visit recording
-		// is unaffected — going back to a channel still updates its
-		// last-visited timestamp.
-		FromHistory bool
-	}
-	MessagesLoadedMsg struct {
-		ChannelID  string
-		Messages   []messages.MessageItem
-		LastReadTS string
-	}
-	OlderMessagesLoadedMsg struct {
-		ChannelID string
-		Messages  []messages.MessageItem
-	}
-	NewMessageMsg struct {
-		ChannelID string
-		Message   messages.MessageItem
-	}
-	SendMessageMsg struct {
-		ChannelID string
-		Text      string
-	}
-	ThreadOpenedMsg struct {
-		ChannelID string
-		ThreadTS  string
-		ParentMsg messages.MessageItem
-	}
-	ThreadRepliesLoadedMsg struct {
-		ThreadTS string
-		Replies  []messages.MessageItem
-	}
-	SendThreadReplyMsg struct {
-		ChannelID string
-		ThreadTS  string
-		Text      string
-	}
-	ThreadReplySentMsg struct {
-		ChannelID string
-		ThreadTS  string
-		LocalTS   string // optimistic-placeholder id; see MessageSentMsg.LocalTS
-		Message   messages.MessageItem
-	}
-	// ThreadReplySendFailedMsg is returned when chat.postMessage for a
-	// thread reply fails. Mirrors MessageSendFailedMsg.
-	ThreadReplySendFailedMsg struct {
-		ChannelID string
-		ThreadTS  string
-		LocalTS   string
-		Reason    string
-	}
-	// ThreadsViewActivatedMsg is dispatched when the user picks the
-	// synthetic Threads sidebar row. The App switches the message pane to
-	// the threads-list view and (re)fetches the involved-threads list.
-	ThreadsViewActivatedMsg struct{}
-	// ThreadsListLoadedMsg carries a freshly loaded list of involved-thread
-	// summaries for the named workspace. The App ignores it if it doesn't
-	// match the active team.
-	ThreadsListLoadedMsg struct {
-		TeamID    string
-		Summaries []cache.ThreadSummary
-		// SubscriptionsAvailable reflects whether the most recent
-		// runSubscriptionPhase succeeded in fetching the authoritative
-		// thread-subscription list. The threads view renders a banner
-		// when false (Task 10 wires the renderer).
-		SubscriptionsAvailable bool
-	}
-	// ThreadsListDirtyMsg is dispatched when something that could affect
-	// the involved-threads list has changed (new message, mention, etc.)
-	// and the list should be refetched. Ignored if not the active team.
-	ThreadsListDirtyMsg struct {
-		TeamID string
-	}
-	ConnectionStateMsg struct {
-		State int // 0=connecting, 1=connected, 2=disconnected
-	}
-	ReactionAddedMsg struct {
-		ChannelID string
-		MessageTS string
-		UserID    string
-		Emoji     string
-	}
-	ReactionRemovedMsg struct {
-		ChannelID string
-		MessageTS string
-		UserID    string
-		Emoji     string
-	}
-	ReactionSentMsg struct {
-		Err error
-	}
-	ChannelMarkedReadMsg struct {
-		ChannelID string
-	}
-	DMNameResolvedMsg struct {
-		ChannelID   string
-		DisplayName string
-		// IsBot is true when the resolved peer is a Slack app or bot.
-		// On first run (before the background users.list fetch has
-		// landed), bot DMs are initially classified as Type="dm" and
-		// the resolveUser path discovers the IsBot flag asynchronously
-		// via users.info. The App handler flips Type to "app" when
-		// this lands so the row hops into the Apps section live.
-		IsBot bool
-	}
-	// UserResolvedMsg arrives asynchronously after main.go's per-workspace
-	// userResolver completes a users.info round-trip for a previously-
-	// unknown message author. The handler patches the in-memory display
-	// name on the messagepane and threadPanel so rows authored by this
-	// user re-render with the real name on the next View().
-	UserResolvedMsg struct {
-		TeamID      string
-		UserID      string
-		DisplayName string
-		IsBot       bool
-	}
-	// UserExternalMsg flags a single user as external (Slack Connect /
-	// shared-channel guest). Emitted by the user-resolution path when a
-	// users.info response shows team_id != workspace TeamID. The App
-	// updates externalUsers and re-pushes user-list state to the pickers.
-	UserExternalMsg struct {
-		UserID     string
-		IsExternal bool
-	}
-	WorkspaceSwitchedMsg struct {
-		TeamID      string
-		TeamName    string
-		Theme       string // resolved theme name (per-workspace or global default)
-		Channels    []sidebar.ChannelItem
-		FinderItems []channelfinder.Item
-		UserNames   map[string]string
-		// ExternalUsers maps userID -> true for users this workspace
-		// considers Slack Connect / shared-channel guests. Hydrated from
-		// cache.User.IsExternal so the mention picker can flag externals
-		// at workspace-switch time, before any new userResolver lookups
-		// fire. Order matters: the App handler applies this BEFORE
-		// SetUserNames so the picker rebuild sees the external flags.
-		ExternalUsers map[string]bool
-		UserID        string
-		CustomEmoji   map[string]string
-		// SectionsProvider supplies Slack-native sidebar sections for this
-		// workspace. Nil means "use config-glob behavior" (the App's
-		// sidebar reverts to its existing name-keyed buckets).
-		SectionsProvider sidebar.SectionsProvider
-	}
-	// ReadStateChangedMsg is sent whenever the persistent read state changes,
-	// so panels that read from cache.GetWorkspaceReadState re-render.
-	// ChannelID may be "" if the change is multi-channel (e.g., batch update
-	// from reconnect catch-up).
-	ReadStateChangedMsg struct {
-		WorkspaceID string
-		ChannelID   string
-	}
-	// ConversationOpenedMsg is sent when Slack delivers an mpim_open,
-	// im_created, group_joined, or channel_joined event. The TeamID
-	// disambiguates events for inactive workspaces; only events whose
-	// TeamID matches the currently-active workspace mutate the live
-	// sidebar — others are persisted in the workspace's WorkspaceContext
-	// for when the user switches in.
-	ConversationOpenedMsg struct {
-		TeamID string
-		Item   sidebar.ChannelItem
-	}
-	// SectionsRefreshedMsg is sent when a workspace's Slack-native
-	// section state has mutated (via channel_section_* WS events) and
-	// the channel list needs to be re-bucketed in the sidebar. Channels
-	// carries a fresh slice with updated Section fields. The App only
-	// rebuckets if TeamID matches the active workspace; inactive
-	// workspaces have already been mutated in-place in their
-	// WorkspaceContext.
-	SectionsRefreshedMsg struct {
-		TeamID   string
-		Channels []sidebar.ChannelItem
-	}
-	// ChannelMembershipMsg is published by membership.Manager when a
-	// channel's member set has been loaded, updated by a delta event, or
-	// refreshed after a TTL miss / reconnect. MemberIDs is the full member
-	// list (not a delta). May be sent for non-active channels — the App
-	// forwards to both compose pickers, which gate on activeChannelID.
-	ChannelMembershipMsg struct {
-		ChannelID string
-		MemberIDs []string
-	}
-	WorkspaceReadyMsg struct {
-		TeamID      string
-		TeamName    string
-		Theme       string // resolved theme name (per-workspace or global default)
-		Channels    []sidebar.ChannelItem
-		FinderItems []channelfinder.Item
-		UserNames   map[string]string
-		// ExternalUsers maps userID -> true for users this workspace
-		// considers Slack Connect / shared-channel guests. Hydrated from
-		// cache.User.IsExternal so the mention picker can flag externals
-		// on startup, before any new userResolver lookups fire. Order
-		// matters: the App handler applies this BEFORE SetUserNames so
-		// the picker rebuild sees the external flags.
-		ExternalUsers map[string]bool
-		UserID        string
-		CustomEmoji   map[string]string
-		// SectionsProvider supplies Slack-native sidebar sections for this
-		// workspace. Nil means "use config-glob behavior" (the App's
-		// sidebar reverts to its existing name-keyed buckets).
-		SectionsProvider sidebar.SectionsProvider
-		// InitialActive is true for exactly one WorkspaceReadyMsg per
-		// program run: the workspace whose team ID matches the configured
-		// default_workspace, or — if no default is configured — the first
-		// workspace to successfully connect. main.go enforces the uniqueness
-		// via sync.Once + atomic router (Task 14). App's handler treats
-		// InitialActive=false as "workspace is up; threads-list kick only".
-		InitialActive bool
-	}
-	// CustomEmojisLoadedMsg is sent when a workspace's custom emoji list
-	// finishes loading in the background, after WorkspaceReadyMsg has
-	// already fired with whatever the goroutine had written so far. App
-	// refreshes the active compose's emoji entry list if this matches the
-	// active workspace.
-	CustomEmojisLoadedMsg struct {
-		TeamID      string
-		CustomEmoji map[string]string
-	}
-	WorkspaceFailedMsg struct {
-		TeamName string
-	}
-	// BrowseableChannelsLoadedMsg is sent after the background fetch of all
-	// public channels (including ones the user has not joined) completes.
-	// The Items have Joined=false; the App merges them into the channel
-	// finder for the matching workspace.
-	BrowseableChannelsLoadedMsg struct {
-		TeamID string
-		Items  []channelfinder.Item
-	}
-	SpinnerTickMsg    struct{}
-	LoadingTimeoutMsg struct{}
-	UserTypingMsg     struct {
-		ChannelID   string
-		UserID      string
-		WorkspaceID string
-	}
-	TypingExpiredMsg struct{}
-	PresenceChangeMsg struct {
-		UserID   string
-		Presence string
-	}
-	// StatusChangeMsg is sent when the authenticated user's own presence
-	// or DND state changes for any workspace. The App routes it to the
-	// status bar only when TeamID matches the active workspace; otherwise
-	// it just updates the App's per-workspace status cache.
-	StatusChangeMsg struct {
-		TeamID     string
-		Presence   string // "active" or "away"; "" means unknown/unchanged
-		DNDEnabled bool
-		DNDEndTS   time.Time
-	}
-	// ToastMsg sets a transient string in the status bar's toast slot. Used
-	// for short error notices (e.g. failed status change). Auto-clears after
-	// 3 seconds via a CopiedClearMsg tick scheduled by the App.
-	ToastMsg struct{ Text string }
-)
-
 type loadingEntry struct {
 	TeamName string
 	Status   string // "connecting", "ready", "failed"
@@ -402,24 +130,6 @@ type dragState struct {
 	// that landed on a message opens that message's thread, mirroring
 	// the Enter keypress. Cleared by clearing the whole dragState.
 	clickedMessage bool
-}
-
-// autoScrollTickMsg is dispatched by tea.Tick while a drag is held near
-// the top or bottom edge of a pane. Each tick scrolls the pane one line
-// in the indicated direction, extends the selection to the new lastY,
-// and (if still at an edge) schedules the next tick. The loop self-
-// terminates when the cursor leaves the edge or the drag ends.
-type autoScrollTickMsg struct{}
-
-// threadFetchDebounceMsg is delivered after the user's threadsview selection
-// stops moving for openThreadDebounceDelay. Carries the (channelID, threadTS,
-// generation) the user had selected at scheduling time; if the App's current
-// pendingThreadFetchGen has advanced past `gen`, the message is dropped — a
-// later j/k has scheduled a fresh fetch.
-type threadFetchDebounceMsg struct {
-	channelID string
-	threadTS  string
-	gen       uint64
 }
 
 // openThreadDebounceDelay is how long openSelectedThreadCmd waits after a
@@ -466,243 +176,6 @@ func boolToInt(b bool) int64 {
 	}
 	return 0
 }
-
-// SwitchWorkspaceFunc is called to switch the active workspace.
-type SwitchWorkspaceFunc func(teamID string) tea.Msg
-
-// ChannelFetchFunc is called when the user selects a channel.
-type ChannelFetchFunc func(channelID, channelName string) tea.Msg
-
-// ChannelCacheReadFunc is called synchronously when the user selects a
-// channel; it returns cached messages from local storage. Returning a
-// non-empty slice causes the messagepane to render immediately without
-// the loading spinner. Returning nil falls through to the network
-// fetcher.
-type ChannelCacheReadFunc func(channelID string) []messages.MessageItem
-
-// OlderMessagesFetchFunc is called when the user scrolls to the top of a channel.
-type OlderMessagesFetchFunc func(channelID, oldestTS string) tea.Msg
-
-// MessageSendFunc is called when the user sends a message. Returns a tea.Msg with the result.
-type MessageSendFunc func(channelID, text string) tea.Msg
-
-// MessageSentMsg is returned after a message is successfully sent.
-// LocalTS, if non-empty, identifies the optimistic placeholder added
-// when the user pressed Enter. The handler uses it to swap the
-// placeholder for the authoritative Message in place. LocalTS may be
-// empty for legacy/test callers that fire this message without a
-// preceding SendMessageMsg.
-type MessageSentMsg struct {
-	ChannelID string
-	LocalTS   string
-	Message   messages.MessageItem
-}
-
-// MessageSendFailedMsg is returned when chat.postMessage fails. The
-// App's handler removes the optimistic placeholder identified by
-// LocalTS (if any) and shows a toast.
-type MessageSendFailedMsg struct {
-	ChannelID string
-	LocalTS   string
-	Reason    string
-}
-
-// EditMessageMsg is emitted when the user submits an edit. App.Update
-// invokes the configured messageEditor and converts the result to
-// MessageEditedMsg.
-type EditMessageMsg struct {
-	ChannelID string
-	TS        string
-	NewText   string
-}
-
-// DeleteMessageMsg is emitted when the user confirms a delete.
-type DeleteMessageMsg struct {
-	ChannelID string
-	TS        string
-}
-
-// MessageEditedMsg carries the result of the chat.update API call.
-type MessageEditedMsg struct {
-	ChannelID string
-	TS        string
-	Err       error
-}
-
-// MessageDeletedMsg carries the result of the chat.delete API call.
-type MessageDeletedMsg struct {
-	ChannelID string
-	TS        string
-	Err       error
-}
-
-// MarkUnreadMsg requests the App to mark the given message as unread.
-// ThreadTS is "" for channel-level mark-unread; non-empty for thread-level
-// (in which case ChannelID is the parent channel and BoundaryTS is the
-// boundary within the thread). BoundaryTS is the ts that should become
-// the new last_read watermark — i.e., the ts of the message immediately
-// before the user's selection. UnreadCount is computed by the dispatcher
-// from the loaded buffer at press time and is forwarded to the sidebar
-// for an exact badge value (0 for thread-level, since the sidebar only
-// tracks channel-level unreads).
-type MarkUnreadMsg struct {
-	ChannelID   string
-	ThreadTS    string
-	BoundaryTS  string
-	UnreadCount int
-}
-
-// MessageMarkedUnreadMsg carries the result of a MarkUnreadFunc call.
-// On success Err is nil and the App's Update arm applies the local
-// state changes (move the unread boundary, update the sidebar badge,
-// flip the threads-view row, emit a toast). On error Err is populated
-// and the toast goes to the failure path; no local state mutates.
-type MessageMarkedUnreadMsg struct {
-	ChannelID   string
-	ThreadTS    string
-	BoundaryTS  string
-	UnreadCount int
-	Err         error
-}
-
-// ChannelMarkedRemoteMsg is dispatched by the WS event handler when
-// Slack pushes a channel_marked / im_marked / group_marked / mpim_marked
-// event (read state changed in another client, or via this client's own
-// mark echoing back). The handler has already persisted the new
-// last_read_ts to SQLite via cache.UpdateChannelReadState; the App's
-// Update arm only updates the UI. No toast.
-type ChannelMarkedRemoteMsg struct {
-	ChannelID   string
-	TS          string
-	UnreadCount int
-}
-
-// ThreadMarkedRemoteMsg is dispatched by the WS event handler when
-// Slack pushes a thread_marked event. Read=true means the thread is
-// now read (clear local boundary + threads-view row); Read=false means
-// it's unread.
-type ThreadMarkedRemoteMsg struct {
-	ChannelID string
-	ThreadTS  string
-	TS        string
-	Read      bool
-}
-
-// WSMessageDeletedMsg is dispatched by the RTM event handler when a
-// message_deleted event arrives. App.Update handles it by removing the
-// message from both panes and the cache.
-type WSMessageDeletedMsg struct {
-	ChannelID string
-	TS        string
-}
-
-// UploadProgressMsg is dispatched out-of-band by the uploader as
-// each file completes. App updates the status-bar toast.
-type UploadProgressMsg struct {
-	Done  int
-	Total int
-}
-
-// UploadResultMsg carries the final result of an upload batch.
-type UploadResultMsg struct {
-	Err error
-}
-
-// UploadFunc performs an upload of one or more files to a channel
-// (with optional thread). It returns a tea.Cmd whose terminal
-// message is UploadResultMsg; intermediate UploadProgressMsg events
-// are dispatched out-of-band via program.Send.
-type UploadFunc func(channelID, threadTS, caption string, attachments []compose.PendingAttachment) tea.Cmd
-
-// MessageEditFunc performs the chat.update API call. Returns a tea.Msg
-// (typically MessageEditedMsg) describing the result.
-type MessageEditFunc func(channelID, ts, newText string) tea.Msg
-
-// MessageDeleteFunc performs the chat.delete API call. Returns a tea.Msg
-// (typically MessageDeletedMsg) describing the result.
-type MessageDeleteFunc func(channelID, ts string) tea.Msg
-
-// MarkUnreadFunc performs the conversations.mark or
-// subscriptions.thread.mark HTTP call (with the rolled-back ts /
-// read=0 form), updates SQLite + in-memory caches if the call
-// succeeded, and returns a tea.Msg (typically MessageMarkedUnreadMsg)
-// describing the result. ThreadTS == "" means channel-level.
-type MarkUnreadFunc func(channelID, threadTS, boundaryTS string, unreadCount int) tea.Msg
-
-// ThreadFetchFunc is called when the user opens a thread.
-type ThreadFetchFunc func(channelID, threadTS string) tea.Msg
-
-// ThreadCacheReadFunc is called synchronously when a thread is opened;
-// returns cached replies (or nil) so the thread panel can populate
-// without waiting for the network. Returning a non-empty slice causes
-// the thread panel to render immediately; the subsequent network
-// response overwrites with authoritative data.
-type ThreadCacheReadFunc func(channelID, threadTS string) []messages.MessageItem
-
-// ThreadMarkFunc is called to mark a thread as read on Slack's servers
-// (subscriptions.thread.mark). channelID is the parent channel, threadTS
-// is the parent message ts, and ts is the latest reply ts the user has now
-// seen. Implementations should be best-effort and non-blocking.
-type ThreadMarkFunc func(channelID, threadTS, ts string)
-
-// ThreadReplySendFunc is called when the user sends a thread reply.
-type ThreadReplySendFunc func(channelID, threadTS, text string) tea.Msg
-
-// ThreadsListFetchFunc loads the involved-threads list for a workspace.
-// Returns the resulting tea.Msg (typically ThreadsListLoadedMsg).
-type ThreadsListFetchFunc func(teamID string) tea.Msg
-
-type ReactionAddFunc func(channelID, messageTS, emoji string) error
-type ReactionRemoveFunc func(channelID, messageTS, emoji string) error
-
-// PermalinkFetchFunc is called to fetch the Slack permalink for a message.
-// For thread replies, pass the reply's ts; Slack returns a thread-aware URL.
-type PermalinkFetchFunc func(ctx context.Context, channelID, ts string) (string, error)
-type FrecentLoadFunc func(limit int) []reactionpicker.EmojiEntry
-type FrecentRecordFunc func(emoji string)
-
-// TypingSendFunc is called to broadcast a typing indicator.
-type TypingSendFunc func(channelID string)
-
-// JoinChannelFunc is called to join a public channel by ID. Returns a tea.Msg
-// describing the result (typically ChannelJoinedMsg or ChannelJoinFailedMsg).
-type JoinChannelFunc func(channelID, channelName string) tea.Msg
-
-// ChannelVisitRecorder is invoked from case ChannelSelectedMsg to let
-// main.go persist the visit (SQLite write + in-memory map update on
-// the WorkspaceContext). Always called regardless of FromHistory.
-type ChannelVisitRecorder func(channelID string)
-
-// ChannelLookupFunc returns metadata for a channel that the App has
-// in its navigation history. Used by navigateBack / navigateForward
-// to skip stale entries (channels the user has left, archived, or
-// kicked from). Returns ok=false when the channel is no longer
-// available in the active workspace.
-type ChannelLookupFunc func(channelID string) (name, channelType string, ok bool)
-
-// ChannelJoinedMsg is sent after the user successfully joins a channel from
-// the channel finder. The App responds by adding the channel to the sidebar
-// (so it appears in the user's regular channel list), marking it as joined in
-// the finder, and switching to it.
-type ChannelJoinedMsg struct {
-	ID   string
-	Name string
-}
-
-// ChannelJoinFailedMsg is sent when the join API call fails.
-type ChannelJoinFailedMsg struct {
-	ID   string
-	Name string
-	Err  error
-}
-
-// clipboardReader abstracts clipboard.Read so tests can inject fake
-// clipboard contents. Production code uses the real clipboard.Read.
-type clipboardReader func(format clipboard.Format) []byte
-
-// defaultClipboardReader is the real clipboard read function. It's
-// overridable per-App via SetClipboardReader for tests.
-var defaultClipboardReader clipboardReader = clipboard.Read
 
 type App struct {
 	// Sub-models
@@ -976,35 +449,6 @@ type App struct {
 	previewTS      string
 	previewAttIdx  int
 }
-
-// previewLoadedMsg is dispatched after a preview thumb has been fetched.
-// The receiver constructs (or, when isCycle is set, swaps the image of)
-// an imgpkg.Preview and stores it on the App.
-type previewLoadedMsg struct {
-	Name         string
-	FileID       string
-	Img          image.Image
-	Path         string
-	SiblingCount int
-	SiblingIndex int
-	// isCycle distinguishes a cycle (h/l/arrow) from a fresh open. On
-	// cycle we mutate the existing overlay rather than constructing a
-	// new one so transient state (e.g. position within wrap-around)
-	// stays coherent.
-	isCycle bool
-}
-
-// previewErrorMsg is dispatched if the preview fetch fails. The error
-// is logged; the overlay is not opened.
-type previewErrorMsg struct {
-	Err error
-}
-
-// previewSpinnerTickMsg drives the loading-state spinner animation.
-// While the preview overlay is open and in its loading state, the
-// Update arm advances the spinner frame and re-schedules another tick;
-// when the image lands (or the overlay closes), the chain stops.
-type previewSpinnerTickMsg struct{}
 
 // previewSpinnerTickInterval is the redraw cadence for the loading
 // spinner. 100ms feels alive without being a CPU hog.
@@ -6203,10 +5647,6 @@ func (a *App) submitEdit(rawValue, translated string) tea.Cmd {
 		}
 	}
 }
-
-// editEmptyToastMsg is delivered when the user tries to submit an
-// edit with empty text.
-type editEmptyToastMsg struct{}
 
 // truncateReason returns s truncated to max characters with an ellipsis.
 // Used for status-bar error toasts.
