+++ internal/ui/callbacks.go
@@ -1,0 +1,124 @@
+// internal/ui/callbacks.go
+//
+// Callback function types used to inject collaborators into App from
+// cmd/slk/main.go. Each Set* method on App takes one of these types
+// and stores it; the App invokes them in response to user actions.
+//
+// Phase 1 of the SOLID refactor of internal/ui/app.go: this file
+// collects every callback type that previously lived in app.go. The
+// callbacks themselves are still flat function pointers — Phase 3
+// will group cohesive subsets into service interfaces (ChannelService,
+// MessageService, ThreadService, ReactionService, WorkspaceService).
+//
+// No semantic change in this commit: same package, same declarations.
+package ui
+
+import (
+	"context"
+
+	tea "charm.land/bubbletea/v2"
+	"golang.design/x/clipboard"
+
+	"github.com/gammons/slk/internal/ui/compose"
+	"github.com/gammons/slk/internal/ui/messages"
+	"github.com/gammons/slk/internal/ui/reactionpicker"
+)
+
+// SwitchWorkspaceFunc is called to switch the active workspace.
+type SwitchWorkspaceFunc func(teamID string) tea.Msg
+
+// ChannelFetchFunc is called when the user selects a channel.
+type ChannelFetchFunc func(channelID, channelName string) tea.Msg
+
+// ChannelCacheReadFunc is called synchronously when the user selects a
+// channel; it returns cached messages from local storage. Returning a
+// non-empty slice causes the messagepane to render immediately without
+// the loading spinner. Returning nil falls through to the network
+// fetcher.
+type ChannelCacheReadFunc func(channelID string) []messages.MessageItem
+
+// OlderMessagesFetchFunc is called when the user scrolls to the top of a channel.
+type OlderMessagesFetchFunc func(channelID, oldestTS string) tea.Msg
+
+// MessageSendFunc is called when the user sends a message. Returns a tea.Msg with the result.
+type MessageSendFunc func(channelID, text string) tea.Msg
+
+// UploadFunc performs an upload of one or more files to a channel
+// (with optional thread). It returns a tea.Cmd whose terminal
+// message is UploadResultMsg; intermediate UploadProgressMsg events
+// are dispatched out-of-band via program.Send.
+type UploadFunc func(channelID, threadTS, caption string, attachments []compose.PendingAttachment) tea.Cmd
+
+// MessageEditFunc performs the chat.update API call. Returns a tea.Msg
+// (typically MessageEditedMsg) describing the result.
+type MessageEditFunc func(channelID, ts, newText string) tea.Msg
+
+// MessageDeleteFunc performs the chat.delete API call. Returns a tea.Msg
+// (typically MessageDeletedMsg) describing the result.
+type MessageDeleteFunc func(channelID, ts string) tea.Msg
+
+// MarkUnreadFunc performs the conversations.mark or
+// subscriptions.thread.mark HTTP call (with the rolled-back ts /
+// read=0 form), updates SQLite + in-memory caches if the call
+// succeeded, and returns a tea.Msg (typically MessageMarkedUnreadMsg)
+// describing the result. ThreadTS == "" means channel-level.
+type MarkUnreadFunc func(channelID, threadTS, boundaryTS string, unreadCount int) tea.Msg
+
+// ThreadFetchFunc is called when the user opens a thread.
+type ThreadFetchFunc func(channelID, threadTS string) tea.Msg
+
+// ThreadCacheReadFunc is called synchronously when a thread is opened;
+// returns cached replies (or nil) so the thread panel can populate
+// without waiting for the network. Returning a non-empty slice causes
+// the thread panel to render immediately; the subsequent network
+// response overwrites with authoritative data.
+type ThreadCacheReadFunc func(channelID, threadTS string) []messages.MessageItem
+
+// ThreadMarkFunc is called to mark a thread as read on Slack's servers
+// (subscriptions.thread.mark). channelID is the parent channel, threadTS
+// is the parent message ts, and ts is the latest reply ts the user has now
+// seen. Implementations should be best-effort and non-blocking.
+type ThreadMarkFunc func(channelID, threadTS, ts string)
+
+// ThreadReplySendFunc is called when the user sends a thread reply.
+type ThreadReplySendFunc func(channelID, threadTS, text string) tea.Msg
+
+// ThreadsListFetchFunc loads the involved-threads list for a workspace.
+// Returns the resulting tea.Msg (typically ThreadsListLoadedMsg).
+type ThreadsListFetchFunc func(teamID string) tea.Msg
+
+type ReactionAddFunc func(channelID, messageTS, emoji string) error
+type ReactionRemoveFunc func(channelID, messageTS, emoji string) error
+
+// PermalinkFetchFunc is called to fetch the Slack permalink for a message.
+// For thread replies, pass the reply's ts; Slack returns a thread-aware URL.
+type PermalinkFetchFunc func(ctx context.Context, channelID, ts string) (string, error)
+type FrecentLoadFunc func(limit int) []reactionpicker.EmojiEntry
+type FrecentRecordFunc func(emoji string)
+
+// TypingSendFunc is called to broadcast a typing indicator.
+type TypingSendFunc func(channelID string)
+
+// JoinChannelFunc is called to join a public channel by ID. Returns a tea.Msg
+// describing the result (typically ChannelJoinedMsg or ChannelJoinFailedMsg).
+type JoinChannelFunc func(channelID, channelName string) tea.Msg
+
+// ChannelVisitRecorder is invoked from case ChannelSelectedMsg to let
+// main.go persist the visit (SQLite write + in-memory map update on
+// the WorkspaceContext). Always called regardless of FromHistory.
+type ChannelVisitRecorder func(channelID string)
+
+// ChannelLookupFunc returns metadata for a channel that the App has
+// in its navigation history. Used by navigateBack / navigateForward
+// to skip stale entries (channels the user has left, archived, or
+// kicked from). Returns ok=false when the channel is no longer
+// available in the active workspace.
+type ChannelLookupFunc func(channelID string) (name, channelType string, ok bool)
+
+// clipboardReader abstracts clipboard.Read so tests can inject fake
+// clipboard contents. Production code uses the real clipboard.Read.
+type clipboardReader func(format clipboard.Format) []byte
+
+// defaultClipboardReader is the real clipboard read function. It's
+// overridable per-App via SetClipboardReader for tests.
+var defaultClipboardReader clipboardReader = clipboard.Read
